Major bug fixes

+++ src/main/java/com/example/library_rest_api/service/BookService.java
@@ -1,5 +1,6 @@
 package com.example.library_rest_api.service;
 
+import com.example.library_rest_api.exception.AccessDeniedException;
 import com.example.library_rest_api.exception.BookWithIsbnAlreadyExists;
 import com.example.library_rest_api.exception.CommonException;
 import com.example.library_rest_api.exception.EntityNotExistException;
@@ -10,6 +11,8 @@ import com.example.library_rest_api.model.BookResponseDto;
 import com.example.library_rest_api.repository.BookRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -26,6 +29,10 @@ public class BookService {
     BookMapper bookMapper;
 
     public BookResponseDto getBookById(Long id) throws CommonException {
+        if (!hasUserRole()) {
+            throw new AccessDeniedException("You do not have permission to update this book.");
+        }
+
         Optional<Book> book = bookRepository.findById(id);
         if (book.isEmpty()) {
             throw new EntityNotExistException("book with id " + id + " not exist");
@@ -35,6 +42,10 @@ public class BookService {
     }
 
     public List<BookResponseDto> getAllBooks() {
+        if (!hasUserRole()) {
+            throw new AccessDeniedException("You do not have permission to update this book.");
+        }
+
         return bookRepository.findAll()
                 .stream()
                 .map(book -> bookMapper.toResponse(book))
@@ -42,6 +53,10 @@ public class BookService {
     }
 
     public BookResponseDto getBookByIsbn(String isbn) throws CommonException {
+        if (!hasUserRole()) {
+            throw new AccessDeniedException("You do not have permission to update this book.");
+        }
+
         Book book = bookRepository.findByIsbn(isbn);
         if (book == null) {
             throw new EntityNotExistException("book with isbn " + isbn + " not exist");
@@ -51,6 +66,10 @@ public class BookService {
     }
 
     public BookResponseDto createBook(BookRequestDto bookRequestDto) {
+        if (!hasAdminRole()) {
+            throw new AccessDeniedException("You do not have permission to update this book.");
+        }
+
         Book book = bookMapper.toEntity(bookRequestDto);
         String isbn = book.getIsbn();
         if (bookRepository.findByIsbn(isbn) != null) {
@@ -61,15 +80,21 @@ public class BookService {
     }
 
     public BookResponseDto updateBook(BookRequestDto book, Long id) throws CommonException {
+        if (!hasAdminRole()) {
+            throw new AccessDeniedException("You do not have permission to update this book.");
+        }
+
         Optional<Book> bookOptional = bookRepository.findById(id);
         if(bookOptional.isPresent()) {
             Book bookToUpdate = bookOptional.get();
+            if (bookRepository.findByIsbn(bookToUpdate.getIsbn()) != null) {
+                throw new BookWithIsbnAlreadyExists("book with isbn " + bookToUpdate.getIsbn() + " already exists");
+            }
             bookToUpdate.setIsbn(book.isbn());
             bookToUpdate.setTitle(book.title());
             bookToUpdate.setAuthor(book.author());
             bookToUpdate.setDescription(book.description());
             bookToUpdate.setGenre(book.genre());
-            bookToUpdate.setAvailable(book.available());
             Book updatedBook = bookRepository.save(bookToUpdate);
             return bookMapper.toResponse(updatedBook);
         }
@@ -77,6 +102,10 @@ public class BookService {
     }
 
     public void deleteBook(Long id) throws CommonException {
+        if (!hasAdminRole()) {
+            throw new AccessDeniedException("You do not have permission to update this book.");
+        }
+
         if(bookRepository.existsById(id))  {
             bookRepository.deleteById(id);
         } else {
@@ -85,6 +114,22 @@ public class BookService {
     }
 
     public void deleteAllBooks() {
+        if (!hasAdminRole()) {
+            throw new AccessDeniedException("You do not have permission to update this book.");
+        }
+
         bookRepository.deleteAll();
+    }
+
+    private boolean hasAdminRole() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        return authentication.getAuthorities().stream()
+                .anyMatch(authority -> authority.getAuthority().equals("ROLE_ADMIN"));
+    }
+
+    private boolean hasUserRole() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        return authentication.getAuthorities().stream()
+                .anyMatch(authority -> authority.getAuthority().equals("ROLE_USER"));
     }
 }
